Add deploy scripts but AA contract not deploying
